🐛(project) fix node permission issues
We must run node with a /home because yarn tries to write to ~/.yarnrc.
If the ID of our host user (with which we run the container) does not
exist in the container (e.g. 1000 exists but 1009 does not exist by
default), then yarn will try to write to "/.yarnrc" at the root of the
system and will fail with a permission error.