🐛(provisioning) fix min count for course video statistics table
The minimum count that can be displayed in the panel of the statistics
table for the course overview table was set 0. For Grafana, a such
setting implies to display results for all statements from the
elasticsearch which does not fit to our expectations as the datasource
contains all courses from the platform. It has been set to 1 so that it
fits to the request based on the course key and only displays data
related to this key.